requirements.txt
anyio==3.6.1
asgiref==3.4.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.2
defusedxml==0.7.1
distlib==0.3.4
dj-database-url==1.0.0
Django==4.0.6
django-filter==21.1
django-heroku==0.3.1
django-mathfilters==1.0.0
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
dnspython==2.2.1
email-validator==1.2.1
fastapi==0.79.0
filelock==3.6.0
gunicorn==20.1.0
h11==0.13.0
httptools==0.4.0
idna==3.3
itsdangerous==2.1.2
itypes==1.2.0
Jinja2==3.1.1
Markdown==3.3.6
MarkupSafe==2.1.1
oauthlib==3.2.0
orjson==3.7.11
platformdirs==2.5.1
psycopg2==2.9.3
pycparser==2.21
pydantic==1.9.1
PyJWT==2.3.0
python-decouple==3.6
python-dotenv==0.20.0
python-multipart==0.0.5
python3-openid==3.2.0
pytz==2022.1
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
sniffio==1.2.0
social-auth-app-django==4.0.0
social-auth-core==4.2.0
sqlparse==0.4.2
starlette==0.19.1
typing_extensions==4.3.0
ujson==5.4.0
uritemplate==4.1.1
urllib3==1.26.9
uvicorn==0.17.6
uvloop==0.16.0
virtualenv==20.14.0
watchfiles==0.16.1
watchgod==0.8.2
websockets==10.3
whitenoise==6.2.0
